Support pythons without yield from, skip tests where not supported
